% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BET_2}
\alias{BET_2}
\title{Alternative, low resilience, Bigeye Tuna base case operating model from recent SS3 Stock Assessment}
\format{
An openMSE object of class 'OM'
}
\source{
Anonymous. 2021. Report of the 2021 Bigeye Stock Assessment Meeting. SCRS/2021/011. Collect. Vol. Sci. Pap. ICCAT, 78(2): 335-485.
}
\usage{
BET_2
}
\description{
Run 6 of the 2021 ICCAT Stock Synthesis 3 assessment (Anonymous 2021) but with steepness = 0.7
}
\examples{
myMSE = runMSE(BET_2)
}
\keyword{datasets}
