% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSH_1}
\alias{BSH_1}
\title{Blue Shark base case operating model from recent SS3 Stock Assessment}
\format{
An openMSE object of class 'OM'
}
\source{
Courtney, D. 2016. Preliminary Stock Synthesis (SS3) Model Runs conducted for North Atlantic blue shark. SCRS/2015/151. Collect. Vol. Sci. Pap. ICCAT, 72(5): 1186-1232.
}
\usage{
BSH_1
}
\description{
Base case Stock Synthesis 3 assessment (Courtney 2016)
}
\examples{
myMSE = runMSE(BSH_1)
}
\keyword{datasets}
