% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWO_MPs.R
\name{CE}
\alias{CE}
\alias{CE_b}
\alias{CE_c}
\alias{CE_d}
\alias{CE_e}
\title{Constant Exploitation with Control Rule}
\usage{
CE(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 1,
  mc = 0.25,
  yrs = c(5, 5),
  ...
)

CE_b(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.834885566858794,
  mc = 0.25,
  yrs = c(5, 5),
  ...
)

CE_c(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.815701097645417,
  mc = 0.25,
  yrs = c(5, 5),
  ...
)

CE_d(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.925689781490413,
  mc = 0.25,
  yrs = c(5, 5),
  ...
)

CE_e(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 1.13785896037804,
  mc = 0.25,
  yrs = c(5, 5),
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{yrs}{numeric values length 2. The year index to calculate the historical mean catches and index,
and the years to smooth the estimate of current exploitation rate}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
This MP aims to keep the exploitation rate at a constant level. The current relative
exploitation rate is calculated as the mean of the catches from 2016:2023 divided
by the mean Combined Index over this same period.
}
\section{Functions}{
\itemize{
\item \code{CE_b()}: Tuned to PGK_short = 0.6 across Reference OMs.

\item \code{CE_c()}: Tuned to PGK_short = 0.7 across Reference OMs.

\item \code{CE_d()}: Tuned to PGK_med = 0.6 across Reference OMs.

\item \code{CE_e()}: Tuned to PGK_long = 0.6 across Reference OMs.

}}
