% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWO_MPs.R
\name{CE2}
\alias{CE2}
\title{Constant Exploitation with Control Rule}
\usage{
CE2(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 1,
  mc = 0.25,
  yrs = c(5, 5),
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{yrs}{numeric values length 2. The year index to calculate the historical mean catches and index,
and the years to smooth the estimate of current exploitation rate}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
This MP aims to keep the exploitation rate at a constant level. The current relative
exploitation rate is calculated as the mean of the catches from 2016:2023 divided
by the mean Combined Index over this same period.
}
