% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_tune.R
\name{CT_1_prep}
\alias{CT_1_prep}
\title{Climate Test Step 1: Spool up historical period of operating models and prepare projections to have zero climate impacts.}
\usage{
CT_1_prep(OM_list)
}
\arguments{
\item{OM_list}{A list of objects of class OM (e.g., BET_1, BET_2, BSH_1, BSH_2)}
}
\value{
A list of the same length of historically reconstructed operating models of class Hist
}
\description{
Prior to tuning MPs, it is necessary to take the set of operating models and remove any existing climate impacts.
}
\examples{
OM_list = list(BET_1,BET_2)       # Create a list of operating models of class 'OM'
Hist_list = CT_1_prep(OM_list)    # Step 1 prep operating models
}
\author{
T. Carruthers
}
