% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_tune.R
\name{CT_2_tune}
\alias{CT_2_tune}
\title{Climate Test Step 2: Tune MPs to a stable biomass outcome.}
\usage{
CT_2_tune(
  Hist_list,
  MPs,
  type = "SSB",
  horizon = 20,
  MP_par_nams = NA,
  MP_par_intervals = NA,
  near_enough = 1e-04,
  tol = 0.005
)
}
\arguments{
\item{Hist_list}{A list of objects of class Hist produced in the first step by CT_1_prep()}

\item{MPs}{A vector of character strings that are the names of MPs.}

\item{type}{Character string. The type of tuning - the default is 'SSB' a stable spawning stock biomass.}

\item{horizon}{Positive integer. The time horizon (number of projected years) at which the outcome (e.g. SSB) is tuned to be the same as current levels.}

\item{MP_par_intervals}{A list (as long as MPs) of vectors, each vector 2 positions long. Optional. These are the lower and upper bounds of the search for the tuning parameter. The algorithm defaults to 1/3 - 3 x the default tuning parameter in each MP.}

\item{near_enough}{Positive real number. Defaults to 1E-4. A measure of whether the tuning function got close enough to the target (e.g. stable SSB after 30 years). Expressed in units of the ratio of SSB(horizon) / SSB(current).}

\item{tol}{Positive real number. the tolerance for convergence of Newton search (optimize). Converged when, between iterations, he tuning parameter changes less than this value.}

\item{MP_par_names}{A vector of character strings as long as MPs - the names of the tuning parameters corresponding to each MP. By default the algorithm assumes the tuning parameters are an MP argument named 'tune'.}
}
\value{
A list of tuned MP functions renamed x_CT.
}
\description{
A function that tunes a set of MPs to achieve current biomass levels in a specified number of years.
}
\examples{
OM_list = list(BET_1,BET_2)                     # Create a list of operating models of class 'OM'
Hist_list = CT_1_prep(OM_list)                  # Step 1 prep operating models
MPs_tuned = CT_2_tune(Hist_list, c("Ir","It"))   # Step 2 tune management procedures
}
\author{
T. Carruthers
}
