% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_results.R
\name{CT_4_summary}
\alias{CT_4_summary}
\title{Climate Test Step 4: Summarize the results of the marginal climate tests.}
\usage{
CT_4_summary(
  data,
  tests = NA,
  MPs = NA,
  RT = 0.9,
  horizon = 30,
  digits = 1,
  grid = F
)
}
\arguments{
\item{data}{A hierarchical list produced in the third step by CT_3_test()}

\item{tests}{A named list of the tests to plot (e.g. c('M', 'K', 'S', 'R', 'C'). Optional, defaults to all tests.}

\item{MPs}{A named list of the MPs to plot (e.g. c('Ir_CT', 'It_CT'). Optional, defaults to all MPs}

\item{RT}{Positive real fraction. Robustness threshold - the fraction of the current SSB.  Used for interpolating to obtain the corresponding percentage level for each MP.}

\item{horizon}{Positive integer. The time horizon (number of projected years) at which the outcome (e.g. SSB) is tuned to be the same as current levels.}

\item{digits}{Positive integer. The number of significant digits for the robustness percentages for each MP on the plot.}

\item{grid}{Boolean. Should gridlines be superimposed on the plot?}
}
\value{
A multipanel figure.
}
\description{
Produces a figure showing the robustness of the various MPs with increasingly stringent climate tests
}
\examples{
OM_list = list(BET_1,BET_2)
Hist_list = CT_1_prep(OM_list)
MPs_tuned = CT_2_tune(Hist_list, c("Ir","It"))
data = CT_3_test(Hist_list, MPs_tuned)
CT_4_summary(data)
}
\author{
T. Carruthers
}
