% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CT_data}
\alias{CT_data}
\title{Example Climate Test Data object}
\format{
A list two positions long, the first is the hierarchical list of MSE objects (openMSE class 'MSE') test x increment, the second is a matrix of the levels of each test.
}
\usage{
CT_data
}
\description{
The output of the function CT_3_test()
}
\examples{
CT_4_summary(CT_data)
}
\keyword{datasets}
