% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_metrics.R
\name{CT_metrics}
\alias{CT_metrics}
\title{Climate Test Metrics summary.}
\usage{
CT_metrics(CT_data, horizon = 20, lev_rnd = 3, dat_rnd = 5)
}
\arguments{
\item{horizon}{Positive integer. The time horizon (number of projected years) at which the outcome (e.g. SSB) is tuned to be the same as current levels.}

\item{lev_rnd}{Positive integer. The number of significant digits to report the levels of each test.}

\item{dat_rnd}{Positive integer. The number of significant digits to report the yield and biomass metrics from each test.}
}
\value{
A list of dataframes
}
\description{
Produces a table of yield and biomass performance at the end of a specified time horizon for each test
}
\examples{
OM_list = list(BET_1,BET_2)
Hist_list = CT_1_prep(OM_list)
MPs_tuned = CT_2_tune(Hist_list, c("Ir","It"))
CT_data = CT_3_test(Hist_list, MPs_tuned)
CT_metrics(CT_data)
}
\author{
T. Carruthers
}
