% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_results.R
\name{CT_proj}
\alias{CT_proj}
\title{Climate Test Metrics HTML summary.}
\usage{
CT_proj(
  CT_data,
  horizon = 30,
  test = NA,
  MPs = NA,
  RT = 0.9,
  rnd = 1,
  denom = 1000,
  unit = "kt",
  fracextra = 0.3,
  CurYr = 2019,
  MPlab_adj = 1.2
)
}
\arguments{
\item{CT_data}{The data object returned from the function CT_3_test.}

\item{horizon}{Positive integer. The time horizon (number of projected years) at which the outcome (e.g. SSB) is tuned to be the same as current levels.}

\item{test}{Character string. The name of the tests you wish to plot results for ("C","M","S","R","K"). Defaults to all tests.}

\item{MPs}{Vector of character strings. The names of the MPs you wish to plot results for. Defaults to all MPs.}

\item{RT}{Fraction. The robustness threshold (fraction of SSB after horizong years).}

\item{rnd}{Positive integer. The number of decimal places to plot test results.}

\item{denom}{Positive real number. A denominator for results (e.g. 1000 if you want to convert kg to tonnes)}

\item{unit}{Character string for units of biomass or yield (e.g. 'kt').}

\item{fracextra}{Positive real number. The amount of white space in which test levels are plotted.}

\item{CurYr}{Positive integer. The current calendar year (last historical year), delineates historical and projected time periods.}

\item{MPlab_adj}{Positive real number. Text adjustment for the plotting of results labels.}
}
\value{
A figure of projected climate test outcomes.
}
\description{
Produces a projection plot of climate test outcomes.
}
\examples{
OM_list = list(BET_1,BET_2)
Hist_list = CT_1_prep(OM_list)
MPs_tuned = CT_2_tune(Hist_list, c("Ir","It"))
CT_data = CT_3_test(Hist_list, MPs_tuned)
CT_proj(CT_data)
}
\author{
T. Carruthers
}
