% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_results.R
\name{CT_tabulate}
\alias{CT_tabulate}
\title{Climate Test Metrics HTML summary.}
\usage{
CT_tabulate(Blist, RT = 0.9, subset = c(1, 2, 3, 4, 5))
}
\arguments{
\item{Blist}{A list of biomass results output produced by CT_metrics.}

\item{RT}{Fraction. The robustness threshold (fraction of SSB after horizong years).}

\item{subset}{Vector of positive integers. The tests you wish to keep in the table.}
}
\value{
A table of values of tests at which the robustness threshold was crossed.
}
\description{
Produces a matrix of performance outcomes (percentage change of each test given robustness threshold) at the end of a specified time horizon for each test
}
\examples{
OM_list = list(BET_1,BET_2)
Hist_list = CT_1_prep(OM_list)
MPs_tuned = CT_2_tune(Hist_list, c("Ir","It"))
CT_data = CT_3_test(Hist_list, MPs_tuned)
results = CT_metrics(CT_data)
CT_tabulate(results$SSB_relative)
}
\author{
T. Carruthers
}
