% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{It}
\alias{It}
\title{Generic Index Target MP for Climate Test Demonstration}
\usage{
It(
  x,
  Data,
  reps = 1,
  tune = 1,
  nyrs = 3,
  maxchng = 0.1,
  maxTACrat = 10,
  calib_yrs = 5
)
}
\arguments{
\item{x}{Simulation number}

\item{Data}{OpenMSE objects of class 'Data'}

\item{reps}{Not used}

\item{tune}{Positive real number, TAC is set by TAC = lastTAC x Recent_index / (recent_historical_index x tune)}

\item{nyrs}{integer, the number of recent years over which to average the index of abundance.}

\item{maxchng}{Positive real number, the maximum rate of change in TAC (upward and downward)}

\item{maxTACrat}{Positive real number, sets the maximum TAC to maxTACrat x mean catches over last historical calib_yrs}

\item{calib_yrs}{Positive integer the number of last historical years to define maxTACrat and recent_historical_index}
}
\description{
Sets TAC advice based on current index levels and a multiplier (targ) of last historical catch/Index
}
\examples{
It(1, Example_datafile)
}
\author{
T. Carruthers
}
