% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWO_MPs.R
\name{MCC9}
\alias{MCC9}
\alias{MCC9_b}
\alias{MCC9_c}
\alias{MCC9_d}
\alias{MCC9_e}
\title{Constant Exploitation with Control Rule}
\usage{
MCC9(x, Data, Data_Lag = 2, Interval = 3, tunepar = 1, mc = NA, ...)

MCC9_b(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.748365885873977,
  mc = NA,
  ...
)

MCC9_c(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.720037451935572,
  mc = NA,
  ...
)

MCC9_d(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.816209907433701,
  mc = NA,
  ...
)

MCC9_e(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.840689429618001,
  mc = NA,
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
goal in this CMP would be to have the catch remain as constant as possible and only increase if the
index rose substantially and only decrease if the index declined substantially. The base TAC (constant
catch) would be 12,600 as this CC TAC that would allow PGK60 and LRP15 (ideally we could get this LRP
down) to be achieved.
}
\section{Functions}{
\itemize{
\item \code{MCC9_b()}: Tuned to PGK_short = 0.6 across Reference OMs.

\item \code{MCC9_c()}: Tuned to PGK_short = 0.7 across Reference OMs.

\item \code{MCC9_d()}: Tuned to PGK_med = 0.6 across Reference OMs.

\item \code{MCC9_e()}: Tuned to PGK_long = 0.6 across Reference OMs.

}}
