% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail_objects.R
\name{objs}
\alias{objs}
\title{What objects of this class are available}
\usage{
objs(classy, package = NULL, msg = TRUE)
}
\arguments{
\item{classy}{A class of object (character string, e.g. 'In', 'OM', 'MSE', 'MP')}

\item{package}{Optional. Names(s) of the package to search for object of class `classy`. String
Default is the `ClimateTest` package.Always searches the global environment as well.}

\item{msg}{Print messages?}
}
\description{
Generic class finder Finds objects of the specified class in the global environment or the
ClimateTest package.
}
\examples{
objs("In", msg=FALSE)
OMs <- objs("OM")
MPs <- objs("MP")
}
\seealso{
\link{avail}
}
\author{
T. Carruthers
}
