% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{Ir}
\alias{Ir}
\title{Generic Index Ratio MP for Climate Test Demonstration}
\usage{
Ir(
  x,
  Data,
  reps = 1,
  tune = 1,
  nyrs = 3,
  maxchng = 0.1,
  maxTACrat = 10,
  calib_yrs = 5
)
}
\arguments{
\item{x}{Simulation number}

\item{Data}{OpenMSE objects of class 'Data'}

\item{reps}{Not used}

\item{tune}{Positive real number, TAC is set by TAC = recent_index x tune x recent_historical_catch/recent_historical_index}

\item{nyrs}{integer, the number of recent years over which to average the index of abundance.}

\item{maxchng}{Positive real number, the maximum rate of change in TAC (upward and downward)}

\item{maxTACrat}{Positive real number, sets the maximum TAC to maxTACrat x mean catches over last historical calib_yrs}

\item{calib_yrs}{Positive integer the number of last historical years to define maxTACrat and the catch/index}
}
\description{
Sets TAC advice based on current index levels and a multiplier (targ) of last historical catch/Index
}
\examples{
Ir(1, Example_datafile)
}
\author{
T. Carruthers
}
